# compute_node/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Samurai Dashboard",
    "description": "REST + WebSocket API для управления роботом.",
    "version": "2.0"
  },
  "paths": {
...
    "/api/v1/fsm/command": {
      "post": {
        "tags": [
          "fsm"
        ],
        "summary": "Send Fsm Command",
        "description": "Послать текстовую команду на Pi (через voice_command топик).\n\nPi-side voice_listener распарсит как обычное голосовое выражение.",
        "operationId": "send_fsm_command_api_v1_fsm_command_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FsmCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fsm/transition": {
      "post": {
        "tags": [
          "fsm"
        ],
        "summary": "Force Fsm Transition",
        "description": "Admin override: принудительно переключить FSM в указанное состояние.",
        "operationId": "force_fsm_transition_api_v1_fsm_transition_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FsmTransitionCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/map/save": {
      "post": {
        "tags": [
          "maps"
        ],
        "summary": "Save Map",
        "description": "Сохранить текущую SLAM карту (ROS2 /map_manager/save). Требует ROS2 stack.",
        "operationId": "save_map_api_v1_map_save_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MapSaveCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/map/load": {
      "post": {
        "tags": [
          "maps"
        ],
        "summary": "Load Map",
        "description": "Загрузить сохранённую карту в SLAM (ROS2 /map_manager/load).",
        "operationId": "load_map_api_v1_map_load_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MapLoadCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommandAck": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          }
        },
        "type": "object",
        "title": "CommandAck",
        "description": "Подтверждение команды (без возвращаемых данных)."
      },
...
      "FsmCommand": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 200,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "FsmCommand",
        "description": "POST /api/fsm/command — текстовая команда (как голосовая)."
      },
...
      "FsmTransitionCommand": {
        "properties": {
          "state": {
            "type": "string",
            "enum": [
              "IDLE",
              "SEARCHING",
              "TARGETING",
              "APPROACHING",
              "GRABBING",
              "CALLING",
              "RETURNING",
              "PATROLLING",
              "FOLLOWING",
              "PATH_REPLAY"
            ],
            "title": "State"
          }
        },
        "type": "object",
        "required": [
          "state"
        ],
        "title": "FsmTransitionCommand",
        "description": "POST /api/fsm/transition — принудительный переход (admin)."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MapLoadCommand": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_\\-]+$",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "MapLoadCommand",
        "description": "POST /api/map/load."
      },
      "MapSaveCommand": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_\\-]+$",
            "title": "Name",
            "description": "Безопасное имя без / и спецсимволов"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "MapSaveCommand",
        "description": "POST /api/map/save."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}